Durchschnitt für Rezeptbewertungen

+++ src/main/java/com/github/learwin/platepalbackend/controller/BewertungController.java
@@ -22,9 +22,11 @@ import java.util.Optional;
 @Controller("/bewertung")
 public class BewertungController {
     private final BewertungRepository bewertungRepository;
+    private final RezeptRepository rezeptRepository;
 
-    BewertungController(BewertungRepository bewertungRepository) {
+    BewertungController(BewertungRepository bewertungRepository, RezeptRepository rezeptRepository) {
         this.bewertungRepository = bewertungRepository;
+        this.rezeptRepository = rezeptRepository;
     }
 
     @Get("/{id}")
@@ -46,6 +48,7 @@ public class BewertungController {
     @ExecuteOn(TaskExecutors.BLOCKING)
     HttpResponse<Bewertung> createBewertung(@Body @Valid Bewertung bewertung) {
         var createdBewerbung = bewertungRepository.save(bewertung);
+        updateMedianBewertung(bewertung);
         return HttpResponse.status(HttpStatus.CREATED).body(createdBewerbung);
     }
 
@@ -53,6 +56,7 @@ public class BewertungController {
     @ExecuteOn(TaskExecutors.BLOCKING)
     HttpResponse<Bewertung> updateBewertung(@Body @Valid Bewertung bewertung) {
         var updatedBewertung = bewertungRepository.update(bewertung);
+        updateMedianBewertung(bewertung);
         return HttpResponse.status(HttpStatus.OK).body(updatedBewertung);
     }
 
@@ -60,7 +64,25 @@ public class BewertungController {
     @ExecuteOn(TaskExecutors.BLOCKING)
     @Status(HttpStatus.NO_CONTENT)
     void deleteBewertung(Long id) {
-        bewertungRepository.deleteById(id);
+        var bewertungOpt = bewertungRepository.findById(id);
+
+        if(bewertungOpt.isPresent()) {
+            var rezeptId = bewertungOpt.get().getRezept_id().getId();
+            bewertungRepository.deleteById(id);
+
+            var durchschnitt = medianBewertung(rezeptId);
+            var rezeptOpt = rezeptRepository.findById(rezeptId);
+
+            if (rezeptOpt.isPresent()) {
+                var rezept = rezeptOpt.get();
+                if (durchschnitt.isPresent())
+                    rezept.setDurchschnittlicheBewertung(durchschnitt.get());
+                else
+                    rezept.setDurchschnittlicheBewertung(null);
+
+                rezeptRepository.update(rezept);
+            }
+        }
     }
 
     @Post(value = "/image/{id}", consumes = {MediaType.MULTIPART_FORM_DATA})
@@ -73,5 +95,20 @@ public class BewertungController {
     @ExecuteOn(TaskExecutors.BLOCKING)
     public HttpResponse<byte[]> getBewertungImage(long id) {
         return ImageHandler.getImageForEntity(id, bewertungRepository, "bewertung");
+    }
+
+    private Optional<Float> medianBewertung(long rezeptId) {
+        return bewertungRepository.getMedianBewertung(rezeptId);
+    }
+
+    private void updateMedianBewertung(Bewertung bewertung) {
+        var durchschnitt  = medianBewertung(bewertung.getRezept_id().getId());
+        var rezeptOpt = rezeptRepository.findById(bewertung.getRezept_id().getId());
+
+        if (rezeptOpt.isPresent()) {
+            var rezept = rezeptOpt.get();
+            rezept.setDurchschnittlicheBewertung(durchschnitt.get());
+            rezeptRepository.update(rezept);
+        }
     }
 }
